+++ oracle/oracle.go
@@ -84,7 +84,7 @@ func New(
 	deviations map[string]sdk.Dec,
 	endpoints map[types.ProviderName]provider.Endpoint,
 	chainConfig bool,
-) (*Oracle, error) {
+) *Oracle {
 	return &Oracle{
 		logger:          logger.With().Str("module", "oracle").Logger(),
 		closer:          pfsync.NewCloser(),
@@ -97,7 +97,7 @@ func New(
 		paramCache:      &ParamCache{params: nil},
 		chainConfig:     chainConfig,
 		endpoints:       endpoints,
-	}, nil
+	}
 }
 
 // LoadProviderPairsAndDeviations loads the on chain pair providers and
@@ -555,63 +555,6 @@ func (o *Oracle) checkCurrencyPairAndDeviations(currentParams, newParams oraclet
 	}
 
 	return nil
-}
-
-// createPairProvidersFromCurrencyPairProvidersList will create the pair providers
-// map used by the price feeder Oracle from a CurrencyPairProvidersList defined by
-// Ojo's oracle module.
-func createPairProvidersFromCurrencyPairProvidersList(
-	currencyPairs oracletypes.CurrencyPairProvidersList,
-) map[types.ProviderName][]types.CurrencyPair {
-	providerPairs := make(map[types.ProviderName][]types.CurrencyPair)
-
-	for _, pair := range currencyPairs {
-		for _, provider := range pair.Providers {
-			if len(pair.PairAddress) > 0 {
-				for _, uniPair := range pair.PairAddress {
-					if (uniPair.AddressProvider == provider) && (uniPair.Address != "") {
-						providerPairs[types.ProviderName(uniPair.AddressProvider)] = append(
-							providerPairs[types.ProviderName(uniPair.AddressProvider)],
-							types.CurrencyPair{
-								Base:    pair.BaseDenom,
-								Quote:   pair.QuoteDenom,
-								Address: uniPair.Address,
-							},
-						)
-					}
-				}
-			} else {
-				providerPairs[types.ProviderName(provider)] = append(
-					providerPairs[types.ProviderName(provider)],
-					types.CurrencyPair{
-						Base:  pair.BaseDenom,
-						Quote: pair.QuoteDenom,
-					},
-				)
-			}
-		}
-	}
-
-	return providerPairs
-}
-
-// createDeviationsFromCurrencyDeviationThresholdList will create the deviations
-// map used by the price feeder Oracle from a CurrencyDeviationThresholdList defined by
-// Ojo's oracle module.
-func createDeviationsFromCurrencyDeviationThresholdList(
-	deviationList oracletypes.CurrencyDeviationThresholdList,
-) (map[string]sdk.Dec, error) {
-	deviations := make(map[string]sdk.Dec, len(deviationList))
-
-	for _, deviation := range deviationList {
-		threshold, err := sdk.NewDecFromStr(deviation.Threshold)
-		if err != nil {
-			return nil, err
-		}
-		deviations[deviation.BaseDenom] = threshold
-	}
-
-	return deviations, nil
 }
 
 func (o *Oracle) tick(ctx context.Context) error {
