+++ oracle/util.go
@@ -6,6 +6,7 @@ import (
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	oracletypes "github.com/ojo-network/ojo/x/oracle/types"
 
 	"github.com/ojo-network/price-feeder/oracle/provider"
 	"github.com/ojo-network/price-feeder/oracle/types"
@@ -221,4 +222,61 @@ func ComputeVwapsByProvider(prices types.AggregatedProviderPrices) types.Currenc
 		vwaps[providerName] = ComputeVWAP(singleProviderCandles)
 	}
 	return vwaps
+}
+
+// createPairProvidersFromCurrencyPairProvidersList will create the pair providers
+// map used by the price feeder Oracle from a CurrencyPairProvidersList defined by
+// Ojo's oracle module.
+func createPairProvidersFromCurrencyPairProvidersList(
+	currencyPairs oracletypes.CurrencyPairProvidersList,
+) map[types.ProviderName][]types.CurrencyPair {
+	providerPairs := make(map[types.ProviderName][]types.CurrencyPair)
+
+	for _, pair := range currencyPairs {
+		for _, provider := range pair.Providers {
+			if len(pair.PairAddress) > 0 {
+				for _, uniPair := range pair.PairAddress {
+					if (uniPair.AddressProvider == provider) && (uniPair.Address != "") {
+						providerPairs[types.ProviderName(uniPair.AddressProvider)] = append(
+							providerPairs[types.ProviderName(uniPair.AddressProvider)],
+							types.CurrencyPair{
+								Base:    pair.BaseDenom,
+								Quote:   pair.QuoteDenom,
+								Address: uniPair.Address,
+							},
+						)
+					}
+				}
+			} else {
+				providerPairs[types.ProviderName(provider)] = append(
+					providerPairs[types.ProviderName(provider)],
+					types.CurrencyPair{
+						Base:  pair.BaseDenom,
+						Quote: pair.QuoteDenom,
+					},
+				)
+			}
+		}
+	}
+
+	return providerPairs
+}
+
+// createDeviationsFromCurrencyDeviationThresholdList will create the deviations
+// map used by the price feeder Oracle from a CurrencyDeviationThresholdList defined by
+// Ojo's oracle module.
+func createDeviationsFromCurrencyDeviationThresholdList(
+	deviationList oracletypes.CurrencyDeviationThresholdList,
+) (map[string]sdk.Dec, error) {
+	deviations := make(map[string]sdk.Dec, len(deviationList))
+
+	for _, deviation := range deviationList {
+		threshold, err := sdk.NewDecFromStr(deviation.Threshold)
+		if err != nil {
+			return nil, err
+		}
+		deviations[deviation.BaseDenom] = threshold
+	}
+
+	return deviations, nil
 }
